Improve encapsulation and add comments
Before background-ing the cat life cycle was done by the Cat instanciator
with the **Future** keyword. Now the this work is done into the Cat.live()
to better encapsulate this function. The user can't be responsible of this.

Add comments to the Cat class in order to put into release branch.